keystone: Allow to disable keystone vhost
In order to minimize keystone downtime we like to keep keystone running
on the non-upgraded nodes for a while after upgrading the first node and
running crowbar_join on it. For that we need to prevent keystone from
starting in the new release on the upgrade node.

This commit introduces a new node attribute (node[:keystone][:disable_vhost])
that will cause the apache_site definition for the keystone vhosts to be
called with "enable false". The attribute is going to be managed by the
Crowbar Upgrade API, which will also make sure of later enabling the new
keystone vhost again.

(cherry picked from commit be63347be0836a5d6fe06c78d7c12f0eb980598d)