nova: Register compute node in nova cell
We need to register compute nodes in nova cells. Otherwise the compute
node is not ready for scheduling VMs on it.

The current setup is racy because we rely on the
[scheduler]"discover_hosts_in_cells_interval" variable (which is set
to 600s currently). So if you add a new compute node, you have to wait
up to 600s until the node is usable.

This problem occured in the CI and also on local deployments. The
nova-conductor.log look like:

nova-conductor.log:2018-11-28 15:26:52.096 29009 ERROR
  nova.conductor.manager No host-to-cell mapping found for selected
  host d52-54-77-77-01-02. Setup is incomplete.: HostMappingNotFound:
  Host 'd52-54-77-77-01-02' is not mapped to any cell

and a bit later the nova-scheduler.log indicates that the host got
discovered:

nova-scheduler.log:2018-11-28 15:27:30.816 28987 INFO
  nova.scheduler.manager Discovered 1 new hosts: cell1:d52-54-77-77-01-02

(cherry picked from commit 718c6172c9eb68a928ad8031c3e041907eb2ead4)