tempest: only assign creator role if needed
Check whether the Barbican barclamp has been applied and only
assign the creator role to the tempest user if it has been.
This role is created by the Barbican barclamp and if the role
does not exist, this role assignment will fail, causing smoke
tests to fail across the board.