galera: Use monitoring user for observing wsrep state
Using the root user with empty password has the downside
that once the root user no longer has an empty password, it
can no longer observe the state. The monitoring user
is more useful for that as it is always available, even
when the cluster was already bootstrapped before. This
is important in case you want to add new nodes to an existing
cluster.