cinder: removed live migrations code
Due to a recent change in cinder, is considered a failure
if the command cinder-manage db online_data_migrations
exits code to 2

docs.openstack.org/releasenotes/cinder/rocky.html

The current tests include live migration and it generates
a failure (error code 2) in ci. So to avoid this problem in a
fresh installation all the migration code is removed. But it
should take in consideration when the cloud is upgraded.