keystone: Fix CA cert Apache config
In 7fdbef8a the cert_required barclamp option was removed, but we still
refer to it in the keystone server recipe. This means it is evaluated as
nil, which causes the Apache CA path to never be set since the wsgi
setup uses crowbar-openstack Apache template[1]. Remove references
to the cert_required parameter, since the CA path will always be
required.

[1] https://github.com/crowbar/crowbar-openstack/blob/release/suse-openstack-cloud/8/chef/cookbooks/crowbar-openstack/templates/default/vhost-wsgi.conf.erb