keystone: Fix path in endpoint update loop
Since the :update_endpoint action of the keystone_register provider
loops over three interfaces, if the path variable is not reset for each
loop it will end up appending new endpoint IDs to the path and resulting
in 404 responses. Use a different variable name within the loop so it
gets reset on each iteration.