keystone: Fix update endpoint for ha (bsc#1114851)
Currently, changing the keystone endpoint protocol does not work in an
HA deployment. The barclamp fails and leaves the deployment in a broken
state with SSL only part way deployed. This is because the values needed
to configure haproxy are determined in the compile phase, and the
resources that enable SSL in haproxy and apache are out of sequence with the
update to the keystone endpoints in the keystone database, so haproxy is
reconfigured and reloaded with the new SSL settings before the endpoint
is changed in keystone and so fall out of alignment.

This commit splits the keystone endpoint update code into its own
recipe, keystone::update_endpoint. It creates a new keystone_register
action, :update_one_endpoint, so that we can update a single endpoint in
the compile phase before haproxy is reloaded. Then a ruby block and sync
marks force apache and haproxy to update their configuration and reload
immediately so that the new internal endpoint can be used to update the
other endpoints.

This causes a small amount of downtime, which is unavoidable and
expected because there is no way to change the keystone endpoints
without also forcing end users to update their auth settings. While
haproxy and apache are reconfiguring themselves, either haproxy or
apache might cause either a gateway error or SSL error if keystone is
accessed, so this adds another wakeup call before the first sync mark to
give the services a chance to come up. Since haproxy is set up in a
roundrobin configuration, using the wakeup call before the sync mark
should in theory ensure the apache service on each controller is ready,
but it's still possible the wakeup calls don't hit each controller, in
which case a retry loop is added to the _update_item method
keystone_register requests to give apache time to come up.

(cherry picked from commit 8dfc0c1c1d3623d35cfb9cb0d80c060c3ad44a41)