Drop "clone_stateless_services" feature
With 5.0 we switch the default of "clone_stateless_services" to false to
have stateless services be managed (restart on failure) by systemd
instead of creating pacemaker resources. When upgrading, existing clouds
will also switch to this configuraton. This patch removes the use of the
"clone_stateless_services" attribute and makes systemd approach the only
supported setup.