nova/neutron: Restart immediately on keystone changes
If the keystone endpoint or keystone credentials change in nova or
neutron's config file, that means the old settings must already be
invalid. If we allow the config file to change but delay restarting the
service, Chef may try to make nova requests, such as listing flavors,
using old keystone settings and will fail.

This change ensures the nova-api and neutron-server services are
restarted immediately when the main config file is changed.

It's possible other services should be doing the same but so far I
haven't observed other services suffering ths ordering problem. There is
always a risk that a chef run will fail in the middle and leave a
service out of sync with its config file, and the service will need to
be restarted manually.

(cherry picked from commit 0aace5337525781f05eb7827d58f3d5dc82103a9)