cinder: add nfs qcow2 option
A new setting is added: nfs_qcow2_volumes. Allows cinder to create volumes
as QCOW2 files rather than raw files for nfs volume.

If QCOW2 is needed, then change it before the creation of nfs backends.
Because this value only could be changed if previously we don't have any nfs backend
to ensure that won't produce any problem with the volumes.

https://docs.openstack.org/cinder/pike/configuration/tables/conf-changes/cinder.html