nova: Use internal glance and neutron endpoints
If not explicitly given, the glance and neutron clients will default to
making requests to the public endpoint for those services. For
service-to-service communication, we always want to use the internal
interface.

This became clear in cases when we are using secure, non-self-signed SSL
certificates for these services. If the certificate CN matches the name
of the internal endpoint but not the public endpoint, the SSL validation
will fail and nova-compute will not be able to retrieve images and
floating IPs.