cinder: Set cinder pool to exclusive by default when using embedded ceph
Pike has been adding support for managing concurrently used pools, so
it spends more than half of its time on assessing the disk usage.
With https://review.openstack.org/#/c/625970/1 this should become better
but until then we can disable the cinder pool size calculation for
embedded ceph, which removes a perceived hang of multiple minutes
when the disk size calculation kicks in.

When using the RBD pool exclusively for Cinder we can now set
`rbd_exclusive_cinder_pool` to `true` and Cinder will use DB information
to calculate provisioned size instead of querying all volumes in the
backend, which will reduce the load on the Ceph cluster and the volume
service.