monasca: Create mon database in Crowbar and use the standard DB
Currently the "mon" database for the Monasca API service is created
via the monasca-installer (ansible based system to deploy Monasca).
To be able to get rid of the monasca-installer code base, do a first
step and create the database (and needed user) the Crowbar way.
While doing that, switch to the standard schema for service databases
and create new attributes in [:monasca][:db_monapi] which are used for
the database. This replaces the attribute
[:monasca][:master][:database_monapi_password].
This also uses the standard database that is setup by Crowbar. So the
database installation done by the monasca-installer is no longer
needed.