nova: Make sure we use the correct conf dir when running nova-manage
By default, if --conf-dir and --conf-file are not specified,
oslo_config will try to load the conf files from the default
locations. While running nova-manage as a non-priviledged user,
this will result in permission denied error even if we are using
a custom conf file in a non-default location. In this case, we
must create a temp conf dir with a temp conf file in it, and we
must specify both --conf-dir and --conf-file options together
in order to avoid loading anything from the default locations.

This patch also improves supportability and resilency by adding
logging and retry capabilities. If we are unable to create
host mapping for whatever reason, we will preserve the log file
for troubleshooting.