monasca: Do extra process/http checks via Crowbar
Instead of using monasca-setup for extra process/http checks, use the
Crowbar LWRP.
monasca-setup detection is currently broken in a strange way[1] and it
requires, that *all* the given processes are running while
monasca-setup is executed. Otherwise, the checks are not
configured (in /etc/monasca/agent/conf.d/processes.yaml). Also it's
possible that only some processes are configured.
This behavior is not wanted. We always want to monitor all the
configured processes and http checks.

[1] https://storyboard.openstack.org/#!/story/2004742