keystone: Remove use of SSLCACertificate
The SSLCACertificate option in Apache is not about specifying the
server's certificate bundle but actually about what certificate chain to
use to validate certificates of client connections. This does not
provide significant value since none of the authentication information
is passed to keystone and the meaning of the parameter has caused
significant user and developer confusion. This change therefore removes
the cert_required and ca_certs parameters in order to clean up the code
and simplify the user interface.