monasca: Deploy elasticsearch-curator with Crowbar
Instead of using monasca-installer for the elasticsearch-curator
setup, use Crowbar.
Also drop the "cron_config" attributes. Cron is no longer
used. Instead, a systemd timer is used and that can be overwritten
with the usual systemd override files.

Note that the elasticsearch-curator setup currently only allows a
single elasticsearch-curator instance (no cluster). That's the same as
with monasca-installer (so no regression).
This is another step in removing monasca-installer.