monasca: Fix grafana config for monasca-persister
The monasca-persister check is done via the process plugin and
it's *not* a http_status check. It was a http_status check when the
Java version of the monasca-persister was used.
So use the same grafana config as we already use for
monasca-notification.

This fixes the Grafana UI when accessing the dashboard node under
"grafana/d/Monasca_dashboard/suse-openstack-cloud-monitoring-monasca"
. The Persister bar there was always black due to the wrong config.