keystone: Add basic osprofiler support
Adding osprofiler[1] support is useful for profiling OpenStack accross
different services. This can be used to find bottlenecks or slow code
inside the different OpenStack services.
osprofiler needs to be enabled in the different services and this
commit starts with keystone.
Once osprofiler is enabled (via the Crowbar attributes), it can be
used like this:

1) Call a command with the openstack client (which supports
osprofiler). Note: You have to know (and use) the HMAC key that is in the
Crowbar attribute stored. If you don't use that HMAC key, osprofiler
won't do anything.

$ openstack --os-profile 9tcwL56PlpAx endpoint list
...
Trace ID: e860c2b0-6945-4767-bbd9-c42c08e6417c

Beside the usual output (in this case the output of "endpoint list"),
there will be a Trace ID written to the command line.
With this Trace ID, the whole trace can be shown with:

$ osprofiler trace show --html --connection-string messaging:// \
    --transport-url \
    "rabbit://openstack:DNUdYtnlMPPY@192.168.192.81:5672/openstack" \
    e860c2b0-6945-4767-bbd9-c42c08e6417c

Note: Currently we use the messaging driver so osprofiler sends it's
trace points to the rabbitmq topic "profiler" which can then be
consumed with the --transport-url.
There are other drivers which might be better (but need to setup other
storage backends) so for now, the messaging driver should be good.

[1] https://docs.openstack.org/osprofiler/latest/