ceilometer: Readd ceilometer user and access (SCRD-7574)
Commit d7601a374d148033 removed the ceilometer API and also the
creation of the ceilometer user and access of this user to the service
tenant. But the ceilometer user is still used by the
ceilometer-polling central agent to get eg. data from Nova.
If the user is not there, the ceilometer/nova_client will emit an
error message:

ceilometer.nova_client Unauthorized: The request you have made \
    requires authentication. (HTTP 401)

This partly reverts commit d7601a374d148033 .