keystone: Make osprofiler connection_string configurable
With the osprofiler connection_string, the backend for storing trace
points can be selected. This is useful when different backends (like
Redis or Elasticsearch) should be used.
It still defaults to "messaging://" as backend which means, that
tracepoints are send to a messaging topic and no special backend is
needed.

(cherry picked from commit 65282ca8f993fbc6fa4895c715724fe379e903b4)