tempest: Conditionally install plugins
If a tempest plugin is installed then by default its tests will be run
unless explicitly turned off in tempest.conf. This breaks our
tempestfull CI because we do not install magnum by default. This change
ensures only plugins for barclamps we have applied will be installed,
and also takes the opportunity to reorganize the list of components
alphabetically and in a git-blame-friendly way and to clean up the
unnecessary special logic for monasca.