glance: run db sync and load metadefs
In bdf0ccef we removed the glance-registry recipe which also removed the
glance-manage db_sync and glance-manage db_load_metadefs executions. We
didn't notice this was a problem at first because it seems that the
glance API service is capable of running the db sync itself, but the
metadef loading was still not happening. This causes tempest failures
because the contents of GET /v2/metadefs/resource_types will be empty.
This patch fixes the issue by adding back this setup steps for the API
service.