tempest: Use existing img_file
We don't ship a cirros disk.img file, but we ship a blank.img file. The
only tests that care about the img_file setting are the barbican tempest
tests and it doesn't matter what type of image is used for those, so
switch to the disk image file that we know will be present.