glance: Set barbican auth endpoint (bsc#1123191)
By default, the castellan key manager library uses
'http://localhost/identity/v3' as its keystone auth endpoint (works in
devstack!). When barbican uploads a signed image to glance, if glance
doesn't override the auth endpoint it will fail to use the key manager.
The symptom of this is a tempest error that looks like:

  TypeError: sendall() argument 1 must be string or buffer, not callable-iterator

which arises from a combination of barbican, glance, and tempest
handling the authentication error very very poorly.