designate: start and run the mdns service
a new recipe for mdns service, it still runs on the same node as api.
Designate uses this internal dns server to perform zone sync between
master.

Without this service, desingate will not be able to perform zone xfers
(AXFR), neither primary or secondary.

rndc key is shared between the dns master node and mdns hosting
node. mdns needs this key to make authoritative zone creation on dns
master

pools.crowbar.yaml file is more like a sample already created for
customers to change as per their requirement. So the pools.crowbar.yaml
file is created on the same node as mdns

The correct ip address in the masters dict of pools file is address of
mdns service where axfr will be requested, using cluster ip in case of
HA