designate: add rndc_secret_key field to proposal
Designate needs an rndc key to CRUD zones on the real bind9 master
server, and updates to bind9 can only be sent via a key that is
mentioned in its(bind9) config.

So the key is supposed to be shared across the nodes, (admin and
designate-worker) therefore adding it to the proposal. One could
rsync the key between the nodes but this opens up different
types of race conditions/synchronization problems.

This key will be used while creating the key-file on admin and
designate-worker node.

The openssl module is part of ruby standard library. And using md5 hexdigest
as that is what rndc expects, a normal 6/8 character password was
deemed invalid during trial runs