nova: Fix discover_hosts group
When the nova-manage command is run, it tries to load configuration from
the default nova config directory even if --config-file is used. The
config files are owned by the root user but readable by the nova group.
If the nova group is not specified in the chef bash resource, the
nova-manage command fails. The failure is hidden and the chef run
continues because the exit code from the bash script comes from the `rm
-f` command the not the nova-manage command. The usual result is that
when the manila server is being booted in preparation for applying the
manila barclamp in CI, it fails. This patch adds the correct group to
the bash resource so that the nova-manage command can complete, and uses
'set -e' so that if it fails for some other reason, it will exit with a
proper failure code and abort the chef run.