tempest: Bump up size of tempest flavor
When a nova server is booted with the 'tempest-stuff' flavor and an
encrypted cinder volume is attached, it can never be detached. Instead,
it will hang in 'detaching' state for a while and then return to
'in-use' with no error messages in the logs. This causes the tempest
tempest.scenario.test_encrypted_cinder_volumes.TestEncryptedCinderVolumes
tests to fail since detaching times out. This change fixes the issue by
bumping up the size of the tempest-stuff flavor that is used for these
tempest tests.