database: apply password update (SCRD-8438)
Currently the create action for mysql database does not check of changed
password.

Without this patch if the admin changes database password, this change
will be applied to the config files however the database is not informed
about this change. As a result that service has looses access to the
database.

This patch relies on the mysql.user table of the database to check of
user existence as well as password hash matching.  Every "input" to
any query is quoted before usage, to avoid any security related error.

An alternate approach to this would be to check if the user connect to
the database with the given credentials, but that requires determining
the correct host string which is not readily available in this context
and will require collecting user specific connection params (including
ssl) and will also require handling of connection close in error as well
as non error cases