memcache: Use first array element as fallback (SCRD-8255)
When this code was refactored, a fallback of node[:memcache][:port] was
used, since the original code in the nova and horizon cookbooks used it.
However, node is not defined in this context, so the fallback causes a
NameError instead. Use the first array element as the fallback instead.

(cherry picked from commit 6da784a4cb173d2023a1be34d4d8ebea12aa74f2)