designate: rely on dns-master entirely
there could be deployments with just single dns-server role in that
there will be no "dns-slaves". In such a case the array append fails as
dnsslaves is nil
So using to_a as fail-safe (nil.to_a == []) way to convert it to array
if not already.

`also_notifies` in such case will have to be empty list
yaml module take case of printing the right thing incase the array is
empty
```yaml
  also_notifies: []
```