neutron: enable designate integration
The config enabled by the flag lets neutron create dns records when
floating ip are created and assigned. (flag: designate_enabled)
The config was previously added but was not enabled as the barclamp was
still under development. Add designate public url to the config.[1]

Set domain name for fixed and floating network.
This allows neutron to assign correct dns domain to the port as they are
created. There is more information at [1]. This is mainly needed when
integrating neutron with external dns service (designate).  Using the neutron
cli to net-update as openstack cli does not support it
(`openstack help network set` has no option for dns-domain)

Fixed a typo in template for neutron as domains are expected to end with
a dot(.)

[1] https://docs.openstack.org/ocata/networking-guide/config-dns-int.html