desginate: Add mdns as hidden master (SOC-9142)
In crowbar's world the dns-master is the master of all and slaves
forward all queries to dns-master. When designate is enabled,
designate-mdns service component(s) become the hidden master(s).

We leave 'also_notifies' empty. It can be configured by the users
via designate-manage utility as and when needed.

designate-mdns service components, in HA, will be
running on multiple hosts and any host can be asked to update a zone
on th real-master. We add the cluster vip as the hidden master in case
HA is enabled otherwise add all servers.