ironic: Use IP also in [swift] section
Swift settings in [glance] already use IP for endpoint URL. This
is used for deployment images.

If config drive is enabled and stored in Swift, this endpoint URL
is not used but instead standard keystone adapter is created based
on [swift] section of Ironic configuration. Added override makes
it possible to use deploy image which can't resolve Crowbar's
internal domain names.