keystone: fix SSL HA scenario (SCRD-9026)
The SSL+HA combo is currently broken (only) in Crowbar SOC7. The
keystone barclamp is failing with the following failure:

FATAL: Errno::ENOENT: ruby_block[synchronize signing keys for founder and remember them for non-HA case] (keystone::server line 347) had an error: Errno::ENOENT: No such file or directory @ rb_sysopen - /etc/keystone/ssl/certs/ca.pem

This happens because there's a code block in the SOC7 keystone
server recipe that is synchronizing the keystone certificates
between the founder and the non-founder nodes. This code block
is a legacy from the days when PKI tokens were still supported [1],
and was re-added [2] after support for PKI tokens was removed [3]
for the only reason that the ceph rados gateway role still requires
these certificates to be present in SOC7 [4].

This commit does the following to remedy the situation:
 - uses the keystone SSL certificate path attributes instead of
 the hard-coded path values. This fixes the CI jobs, because the
 CI now explicitly configures the CA certificate attribute in the
 keystone barclamp to point to the CA certificate installed with
 the SUSE `ca-certificates' package [5]
 - remove the code that synchronizes the certificate file contents
 between the founder and non-founder nodes, because this is no
 longer required. The only thing that is required by the ceph
 barclamp is that the certificate file contents be saved as node
 attributes [6], to be converted to nss db format [7]. With the
 synchronization code out of the way, the sync marks also become
 useless and are removed.

[1] https://github.com/crowbar/crowbar-openstack/blob/51bbfcf0a613c5aca975d32b8f83b6102b0d4623/chef/cookbooks/keystone/recipes/server.rb#L311-L380
[2] https://github.com/crowbar/crowbar-openstack/pull/743
[3] https://github.com/crowbar/crowbar-openstack/pull/702/commits/da320389b2591b44e4088dbcdcb164998cabaf47
[4] https://github.com/crowbar/crowbar-ceph/pull/83
[5] https://github.com/SUSE-Cloud/automation/commit/4194b0ee6fbe05305d17bca2166ca87fd5cefc02
[6] https://github.com/crowbar/crowbar-ceph/blob/stable/4.0/chef/cookbooks/ceph/recipes/radosgw_keystone.rb#L112-L139
[7] http://docs.ceph.com/docs/jewel/radosgw/keystone/