neutron: use crm_resource restart for restarting neutron-l3-ha-service
When we restart via systemd, it can happen that the service already
crashed and then 'service neutron-l3-ha-service status" will (correctly)
report that the service isn't running. And then the restart is not being
executed due to:

  INFO: Ignoring restart action for neutron-l3-ha-service service since not running on this node (d52-54-77-77-01-01)

which will then later cause a pacemaker failcount. the pacemaker
resource restart hopefully does not have this problem.