nova: Show stdout/stderr when "openstack flavor list" fails
When running into the problem that the "openstack flavor list" command
fails, we want to know the error and not only see a comment that the
nova api might be down.
This is useful for debugging.

(cherry picked from commit b8776b650f511dc9b42e0e579dc6f49de590a690)