neutron: Stop creating / watching .openrc
On a toggle of the keystone services configuration, the .openrc
is written too early and neutron-l3-ha-service is restarted
prior the configuration yaml being updated. So given that
the l3-ha-service only fetches information from the yaml, we don't
need to trigger restart and we don't need to write the .openrc.