monasca: move Grafana DB creation
This commit moves Grafana DB creation to the Horizon barclamp
since this is where the Grafana service is being deployed.

This one is crucial for the upgrade since a controller only
upgrade does not necessarily apply the Monasca barclamp
(unless the monasca-server role is assigned to a controller).
Consequently, the grafana database (which used to be in the
MariaDB instance on the Monasca node) won't be created in the
regular MariaDB instance, causing the chef run to fail.