swift: Sync HA nodes (SOC-9683)
In most barclamps using syncmarks existing syncmarks are reset when applying
proposal. In Swift barclamp, swift-proxy is the only "HA" role and it is not
executed as the first one (swift-storage and swift-ring-compute need to be
run first). Because of that, first chef-client run during proposal commit
might not include the founder node and syncmark mechanism would fail in all
cookbooks.

Because syncmarks are not reset when applying Swift proposal, nodes can be
out of sync beyond normal syncmark timeout. This 'sync' syncmark should
line them up and correct any time offsets.