designate: Use server node for VIP look ups (SOC-9631)
If designate is deployed across two clusters, one for the server and the
other for the workers, when the worker recipe applies, it will use the
current node to look up the VIP it should use to talk to the server,
which will fail, since it doesn't have one. Since we can guarantee a
server role exists, make use of the first server role node for those
look ups.